classmint-teacher/requirements.txt
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-CORS==6.0.1
qrcode[pil]==7.4.2
Pillow==10.0.1
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3
